added authentication to notes (user notes)

+++ backend/src/controllers/notesControllers.ts
@@ -8,10 +8,15 @@ import {
 } from "../utils/noteInterface";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertsIsDefined } from "../utils/assersDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticationUserId = req.session.userId
   try {
-    const notes = await NoteModel.find().exec();
+    //check if userId is null
+    assertsIsDefined(authenticationUserId)
+    // gets only auth user notes
+    const notes = await NoteModel.find({userId: authenticationUserId}).exec();
     res.status(200).json(notes);
     res.end();
   } catch (error) {
@@ -21,8 +26,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId; // params are after "/:"
+  const authenticationUserId = req.session.userId
 
   try {
+    assertsIsDefined(authenticationUserId)
     if (!mongoose.isValidObjectId(noteId)) {
       // checks if id suits for id format
       throw createHttpError(400, "invalid note Id"); //bad request
@@ -31,6 +38,11 @@ export const getNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, "Note not found");
     }
+
+    if (!note.userId?.equals(authenticationUserId)) {
+      throw createHttpError(401,"You cannot access this note")
+    }
+
     res.status(200).json(note); // back to client
   } catch (error) {
     next(error);
@@ -45,14 +57,18 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticationUserId = req.session.userId
 
   try {
+    assertsIsDefined(authenticationUserId)
+
     if (!title) {
       // We can declare how we handle if no title. Normally it throws error, if there is no title in request. We said that title is required is schema
       throw createHttpError(400, "title is needed"); // bad request, missing param in request
     }
     const newNote = await NoteModel.create({
       //Promise
+      userId: authenticationUserId,
       title: title,
       text: text,
     });
@@ -70,7 +86,10 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticationUserId = req.session.userId
+
   try {
+    assertsIsDefined(authenticationUserId)
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "invalid note Id"); 
     }
@@ -82,6 +101,11 @@ export const updateNote: RequestHandler<
     if (!note) {
       throw createHttpError(404, "Note not found");
     }
+
+    if (!note.userId?.equals(authenticationUserId)) {
+      throw createHttpError(401,"You cannot access this note")
+    }
+
     note.title = newTitle;
     note.text = newText;
 
@@ -93,7 +117,10 @@ export const updateNote: RequestHandler<
 };
 export const deleteNote: RequestHandler <DeleteNoteParamas, unknown, unknown> = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticationUserId = req.session.userId
   try {
+
+    assertsIsDefined(authenticationUserId)
 
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400,"Invalid note Id")
@@ -103,6 +130,11 @@ export const deleteNote: RequestHandler <DeleteNoteParamas, unknown, unknown> = 
     if (!note) {
       throw createHttpError(404,"Note not found")
     }
+
+    if (!note.userId?.equals(authenticationUserId)) {
+      throw createHttpError(401,"You cannot access this note")
+    }
+    
     await NoteModel.findByIdAndDelete(noteId);
     res.sendStatus(204);
